Trees for genotypes A-D
I ran timetrees as an experiment but the results are very poor and so
we are only using divergence trees. This is expected given the large
amount of recombination present which we haven't attempted to detect/
remove. Perhaps it's possible within a subgenotype, but even then I
don't think so.

The rooting isn't quite right -- probably due to LBA -- which I'll attend
to in a subsequent commit.